feat: improved PreventNetherRoof
The player is not getting teleported under the bedrock layer now, if the blocks are Netherrack they are getting replaced with air and the player is getting teleported to that spot.
If the blocks are not Netherrack, the player gets teleported into these blocks to prevent removing blocks that shouldn't be removed.